# docs/openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "zentry API Docs",
		"description": "zentry API Docs",
		"contact": {
			"email": "deshraj@zentry.gg/placeholder"
		},
		"license": {
			"name": "Apache 2.0"
		},
		"version": "v1"
	},
	"servers": [
		{
			"url": "https://api.zentry.gg/placeholder/"
		}
	],
	"security": [
		{
			"ApiKeyAuth": [

			]
		}
	],
	"paths": {
...
		"/v1/memories/": {
			"get": {
				"tags": [
					"memories"
				],
				"description": "Get all memories",
				"operationId": "memories_list",
				"parameters": [
					{
						"name": "user_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by user ID"
					},
					{
						"name": "agent_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by agent ID"
					},
					{
						"name": "app_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by app ID"
					},
					{
						"name": "run_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by run ID"
					},
					{
						"name": "metadata",
						"in": "query",
						"schema": {
							"type": "object"
						},
						"description": "Filter memories by metadata (JSON string)",
						"style": "deepObject",
						"explode": true
					},
					{
						"name": "categories",
						"in": "query",
						"schema": {
							"type": "array",
							"items": { "type": "string" }
						},
						"description": "Filter memories by categories"
					},
					{
						"name": "org_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by organization ID."
					},
					{
						"name": "project_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by project ID."
					},
					{
						"name":"fields",
						"in": "query",
						"schema": { "type": "array", "items": { "type": "string" } },
						"description": "Filter memories by fields"
					},
					{
						"name":"keywords",
						"in": "query",
						"schema": { "type": "string" },
						"description": "Filter memories by keywords"
					},
					{
						"name": "page",
						"in": "query",
						"schema": { "type": "integer" },
						"description": "Page number for pagination. Default: 1"
					},
					{
						"name": "page_size",
						"in": "query",
						"schema": { "type": "integer" },
						"description": "Number of items per page. Default: 100"
					},
					{
						"name": "start_date",
						"in": "query",
						"schema": { "type": "string" },
						"description": "Filter memories by start date"
					},
					{
						"name": "end_date",
						"in": "query",
						"schema": { "type": "string" },
						"description": "Filter memories by end date"
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "string"
											},
											"name": {
												"type": "string"
											},
											"input": {
												"type": "array",
												"items": {
													"type": "object",
													"properties": {
														"role": {
															"type": "string"
														},
														"content": {
															"type": "string"
														}
													}
												}
											},
											"created_at": {
												"type": "string",
												"format": "date-time"
											},
											"updated_at": {
												"type": "string",
												"format": "date-time"
											},
											"owner": {
												"type": "string"
											},
											"immutable": {
												"type": "boolean",
												"description": "Whether the memory is immutable.",
												"title": "Immutable",
												"default": false
											},
											"expiration_date": {
												"type": "string",
												"format": "date-time",
												"description": "The date and time when the memory will expire. Format: YYYY-MM-DD",
												"title": "Expiration date",
												"nullable": true,
												"default": null
											},
											"organization": {
												"type": "string"
											},
											"metadata": {
												"type": "object"
											}
										},
										"required": [
											"id",
											"name",
											"created_at",
											"updated_at",
											"total_memories",
											"owner",
											"organization",
											"type"
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "One of the filters: app_id, user_id, agent_id, run_id is required!"
										}
									}
								}
							}
						}
					}
				},
				"x-code-samples": [
				{
				"lang": "Python",
				"source": "# To use the Python SDK, install the package:\n# pip install zentryai\n\nfrom zentry import MemoryClient\nclient = MemoryClient(api_key=\"your_api_key\", org_id=\"your_org_id\", project_id=\"your_project_id\")\n\n# Retrieve memories for a specific user\nuser_memories = client.get_all(user_id=\"<user_id>\")\n\nprint(user_memories)"
				},
				{
				"lang": "JavaScript",
				"source": "// To use the JavaScript SDK, install the package:\n// npm i zentryai\n\nimport MemoryClient from 'zentryai';\nconst client = new MemoryClient({ apiKey: \"your-api-key\" });\n\n// Retrieve memories for a specific user\nclient.getAll({ user_id: \"<user_id>\" })\n  .then(result => console.log(result))\n  .catch(error => console.error(error));"
				},
				{
				"lang": "cURL",
				"source": "curl --location --request GET 'https://api.zentry.gg/v1/memories/' \\\n--header 'Authorization: Token <api-key>'"
				},
				{
				"lang": "Go",
				"source": "package main\n\nimport (\n\t\"fmt\"\n\t\"net/http\"\n\t\"io/ioutil\"\n)\n\nfunc main() {\n\n\turl := \"https://api.zentry.gg/v1/memories/\"\n\n\treq, _ := http.NewRequest(\"GET\", url, nil)\n\n\treq.Header.Add(\"Authorization\", \"Token <api-key>\")\n\n\tres, _ := http.DefaultClient.Do(req)\n\n\tdefer res.Body.Close()\n\tbody, _ := ioutil.ReadAll(res.Body)\n\n\tfmt.Println(res)\n\tfmt.Println(string(body))\n\n}"
				},
				{
				"lang": "PHP",
				"source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, [\n  CURLOPT_URL => \"https://api.zentry.gg/v1/memories/\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"GET\",\n  CURLOPT_HTTPHEADER => [\n    \"Authorization: Token <api-key>\"\n  ],\n]);\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
				},
				{
				"lang": "Java",
				"source": "HttpResponse<String> response = Unirest.get(\"https://api.zentry.gg/v1/memories/\")\n  .header(\"Authorization\", \"Token <api-key>\")\n  .asString();"
					}
				]
			},
			"post": {
				"tags": [
					"memories"
				],
				"description": "Add memories",
				"operationId": "memories_create",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemoryInput"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful memory creation",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "string"
											},
											"data": {
												"type": "object",
												"properties": {
													"memory": {
														"type": "string"
													}
												},
												"required": [
													"memory"
												]
											},
											"event": {
												"type": "string",
												"enum": [
													"ADD",
													"UPDATE",
													"DELETE"
												]
											}
										},
										"required": [
											"id",
											"data",
											"event"
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request. Invalid input data. Please refer to the memory creation documentation at https://docs.zentry.gg/platform/quickstart#4-1-create-memories for correct formatting and required fields.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"error",
										"details"
									],
									"example": {
										"error": "400 Bad Request",
										"details": {
											"message": "Invalid input data. Please refer to the memory creation documentation at https://docs.zentry.gg/platform/quickstart#4-1-create-memories for correct formatting and required fields."
										}
									}
								}
							}
						}
					}
				},
				"x-code-samples": [
				{
				"lang": "Python",
				"source": "# To use the Python SDK, install the package:\n# pip install zentryai\n\nfrom zentry import MemoryClient\n\nclient = MemoryClient(api_key=\"your_api_key\", org_id=\"your_org_id\", project_id=\"your_project_id\")\n\nmessages = [\n    {\"role\": \"user\", \"content\": \"<user-message>\"},\n    {\"role\": \"assistant\", \"content\": \"<assistant-response>\"}\n]\n\nclient.add(messages, user_id=\"<user-id>\", version=\"v2\")"
				},
				{
				"lang": "JavaScript",
				"source": "// To use the JavaScript SDK, install the package:\n// npm i zentryai\n\nimport MemoryClient from 'zentryai';\nconst client = new MemoryClient({ apiKey: \"your-api-key\" });\n\nconst messages = [\n  { role: \"user\", content: \"Hi, I'm Alex. I'm a vegetarian and I'm allergic to nuts.\" },\n  { role: \"assistant\", content: \"Hello Alex! I've noted that you're a vegetarian and have a nut allergy. I'll keep this in mind for any food-related recommendations or discussions.\" }\n];\n\nclient.add(messages, { user_id: \"<user_id>\", version: \"v2\" })\n  .then(result => console.log(result))\n  .catch(error => console.error(error));"
				},				  
				{
				"lang": "cURL",
				"source": "curl --request POST \\\n  --url https://api.zentry.gg/v1/memories/ \\\n  --header 'Authorization: Token <api-key>' \\\n  --header 'Content-Type: application/json' \\\n  --data '{\n  \"messages\": [\n    {}\n  ],\n  \"agent_id\": \"<string>\",\n  \"user_id\": \"<string>\",\n  \"app_id\": \"<string>\",\n  \"run_id\": \"<string>\",\n  \"metadata\": {},\n  \"includes\": \"<string>\",\n  \"excludes\": \"<string>\",\n  \"infer\": true,\n  \"custom_categories\": {}, \n  \"org_id\": \"<string>\",\n  \"project_id\": \"<string>\",\n  \"version\": \"v2\"\n}'"
				},
				{
				"lang": "Go",
				"source": "package main\n\nimport (\n\t\"fmt\"\n\t\"strings\"\n\t\"net/http\"\n\t\"io/ioutil\"\n)\n\nfunc main() {\n\n\turl := \"https://api.zentry.gg/v1/memories/\"\n\n\tpayload := strings.NewReader(\"{\n  \\\"messages\\\": [\n    {}\n  ],\n  \\\"agent_id\\\": \\\"<string>\\\",\n  \\\"user_id\\\": \\\"<string>\\\",\n  \\\"app_id\\\": \\\"<string>\\\",\n  \\\"run_id\\\": \\\"<string>\\\",\n  \\\"metadata\\\": {},\n  \\\"includes\\\": \\\"<string>\\\",\n  \\\"excludes\\\": \\\"<string>\\\",\n  \\\"infer\\\": true,\n  \\\"custom_categories\\\": {},\n  \\\"org_id\\\": \\\"<string>\\\",\n  \\\"project_id\\\": \\\"<string>\",\n  \\\"version\\\": \"v2\"\n}\")\n\n\treq, _ := http.NewRequest(\"POST\", url, payload)\n\n\treq.Header.Add(\"Authorization\", \"Token <api-key>\")\n\treq.Header.Add(\"Content-Type\", \"application/json\")\n\n\tres, _ := http.DefaultClient.Do(req)\n\n\tdefer res.Body.Close()\n\tbody, _ := ioutil.ReadAll(res.Body)\n\n\tfmt.Println(res)\n\tfmt.Println(string(body))\n\n}"
				},
				{
				"lang": "PHP",
				"source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, [\n  CURLOPT_URL => \"https://api.zentry.gg/v1/memories/\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"POST\",\n  CURLOPT_POSTFIELDS => \"{\n  \\\"messages\\\": [\n    {}\n  ],\n  \\\"agent_id\\\": \\\"<string>\\\",\n  \\\"user_id\\\": \\\"<string>\\\",\n  \\\"app_id\\\": \\\"<string>\\\",\n  \\\"run_id\\\": \\\"<string>\\\",\n  \\\"metadata\\\": {},\n  \\\"includes\\\": \\\"<string>\\\",\n  \\\"excludes\\\": \\\"<string>\\\",\n  \\\"infer\\\": true,\n  \\\"custom_categories\\\": {}, \n  \\\"org_id\\\": \\\"<string>\\\",\n  \\\"project_id\\\": \\\"<string>\",\n  \\\"version\\\": \"v2\"\n}\",\n  CURLOPT_HTTPHEADER => [\n    \"Authorization: Token <api-key>\",\n    \"Content-Type: application/json\"\n  ],\n]);\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
				},
				{
				"lang": "Java",
				"source": "HttpResponse<String> response = Unirest.post(\"https://api.zentry.gg/v1/memories/\")\n  .header(\"Authorization\", \"Token <api-key>\")\n  .header(\"Content-Type\", \"application/json\")\n  .body(\"{\n  \\\"messages\\\": [\n    {}\n  ],\n  \\\"agent_id\\\": \\\"<string>\\\",\n  \\\"user_id\\\": \\\"<string>\\\",\n  \\\"app_id\\\": \\\"<string>\\\",\n  \\\"run_id\\\": \\\"<string>\\\",\n  \\\"metadata\\\": {},\n  \\\"includes\\\": \\\"<string>\\\",\n  \\\"excludes\\\": \\\"<string>\\\",\n  \\\"infer\\\": true,\n  \\\"custom_categories\\\": {}, \n  \\\"org_id\\\": \\\"<string>\\\",\n  \\\"project_id\\\": \\\"<string>\",\n  \\\"version\\\": \"v2\"\n}\")\n  .asString();"
					}
				],
				"x-codegen-request-body-name": "data"
			},
			"delete": {
				"tags": [
					"memories"
				],
				"description": "Delete memories",
				"operationId": "memories_delete",
				"parameters": [
					{
						"name": "user_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by user ID"
					},
					{
						"name": "agent_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by agent ID"
					},
					{
						"name": "app_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by app ID"
					},
					{
						"name": "run_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by run ID"
					},
					{
						"name": "metadata",
						"in": "query",
						"schema": {
							"type": "object"
						},
						"description": "Filter memories by metadata (JSON string)",
						"style": "deepObject",
						"explode": true
					},
					{
						"name": "org_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by organization ID."
					},
					{
						"name": "project_id",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter memories by project ID."
					}
				],
				"responses": {
					"204": {
						"description": "Successful deletion of memories",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Memories deleted successfully!"
										}
									}
								}
							}
						}
					}
				},
				"x-code-samples": [
				{
				"lang": "Python",
				"source": "# To use the Python SDK, install the package:\n# pip install zentryai\n\nfrom zentry import MemoryClient\nclient = MemoryClient(api_key=\"your_api_key\", org_id=\"your_org_id\", project_id=\"your_project_id\")\n\n# Delete all memories for a specific user\nclient.delete_all(user_id=\"<user_id>\")"
				},
				{
				"lang": "JavaScript",
				"source": "// To use the JavaScript SDK, install the package:\n// npm i zentryai\n\nimport MemoryClient from 'zentryai';\nconst client = new MemoryClient({ apiKey: \"your-api-key\" });\n\n// Delete all memories for a specific user\nclient.deleteAll({ user_id: \"<user_id>\" })\n  .then(result => console.log(result))\n  .catch(error => console.error(error));"
				},				  
				{
				"lang": "cURL",
				"source": "curl --request DELETE \\\n  --url https://api.zentry.gg/v1/memories/ \\\n  --header 'Authorization: Token <api-key>'"
				},
				{
				"lang": "Go",
				"source": "package main\n\nimport (\n\t\"fmt\"\n\t\"net/http\"\n\t\"io/ioutil\"\n)\n\nfunc main() {\n\n\turl := \"https://api.zentry.gg/v1/memories/\"\n\n\treq, _ := http.NewRequest(\"DELETE\", url, nil)\n\n\treq.Header.Add(\"Authorization\", \"Token <api-key>\")\n\n\tres, _ := http.DefaultClient.Do(req)\n\n\tdefer res.Body.Close()\n\tbody, _ := ioutil.ReadAll(res.Body)\n\n\tfmt.Println(res)\n\tfmt.Println(string(body))\n\n}"
				},
				{
				"lang": "PHP",
				"source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, [\n  CURLOPT_URL => \"https://api.zentry.gg/v1/memories/\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"DELETE\",\n  CURLOPT_HTTPHEADER => [\n    \"Authorization: Token <api-key>\"\n  ],\n]);\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
				},
				{
				"lang": "Java",
				"source": "HttpResponse<String> response = Unirest.delete(\"https://api.zentry.gg/v1/memories/\")\n  .header(\"Authorization\", \"Token <api-key>\")\n  .asString();"
					}
				],
				"x-codegen-request-body-name": "data"
			}
		},
...
		"/v1/memories/search/": {
			"post": {
				"tags": [
					"memories"
				],
				"description": "Perform a semantic search on memories.",
				"operationId": "memories_search_create",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemorySearchInput"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "string",
												"format": "uuid",
												"description": "Unique identifier for the memory"
											},
											"memory": {
												"type": "string",
												"description": "The content of the memory"
											},
											"user_id": {
												"type": "string",
												"description": "The identifier of the user associated with this memory"
											},
											"metadata": {
												"type": "object",
												"nullable": true,
												"description": "Additional metadata associated with the memory"
											},
											"categories": {
												"type": "array",
												"items": {
													"type": "string"
												},
												"description": "Categories associated with the memory"
											},
											"immutable": {
												"type": "boolean",
												"description": "Whether the memory is immutable.",
												"title": "Immutable",
												"default": false
											},
											"expiration_date": {
												"type": "string",
												"format": "date-time",
												"description": "The date and time when the memory will expire. Format: YYYY-MM-DD",
												"title": "Expiration date",
												"nullable": true,
												"default": null
											},
											"created_at": {
												"type": "string",
												"format": "date-time",
												"description": "The timestamp when the memory was created"
											},
											"updated_at": {
												"type": "string",
												"format": "date-time",
												"description": "The timestamp when the memory was last updated"
											}
										},
										"required": [
											"id",
											"memory",
											"user_id",
											"created_at",
											"updated_at"
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "At least one of the filters: agent_id, user_id, app_id, run_id is required!"
										}
									}
								}
							}
						}
					}
				},
				"x-code-samples": [
				{
				"lang": "Python",
				"source": "# To use the Python SDK, install the package:\n# pip install zentryai\n\nfrom zentry import MemoryClient\nclient = MemoryClient(api_key=\"your_api_key\", org_id=\"your_org_id\", project_id=\"your_project_id\")\n\nquery = \"Your search query here\"\n\nresults = client.search(query, user_id=\"<user_id>\", output_format=\"v1.1\")\nprint(results)"
				},
				{
				"lang": "JavaScript",
				"source": "// To use the JavaScript SDK, install the package:\n// npm i zentryai\n\nimport MemoryClient from 'zentryai';\nconst client = new MemoryClient({ apiKey: \"your-api-key\" });\n\nconst query = \"Your search query here\";\n\nclient.search(query, { user_id: \"<user_id>\", output_format: \"v1.1\" })\n  .then(result => console.log(result))\n  .catch(error => console.error(error));"
				},				  
				{
				"lang": "cURL",
				"source": "curl --request POST \\\n  --url https://api.zentry.gg/v1/memories/search/ \\\n  --header 'Authorization: Token <api-key>' \\\n  --header 'Content-Type: application/json' \\\n  --data '{\n  \"query\": \"<string>\",\n  \"agent_id\": \"<string>\",\n  \"user_id\": \"<string>\",\n  \"app_id\": \"<string>\",\n  \"run_id\": \"<string>\",\n  \"metadata\": {},\n  \"top_k\": 123,\n  \"fields\": [\n    \"<string>\"\n  ],\n  \"rerank\": true,\n  \"org_id\": \"<string>\",\n  \"project_id\": \"<string>\"\n}'"
				},
				{
				"lang": "Go",
				"source": "package main\n\nimport (\n\t\"fmt\"\n\t\"net/http\"\n\t\"io/ioutil\"\n)\n\nfunc main() {\n\n\turl := \"https://api.zentry.gg/v1/memories/\"\n\n\treq, _ := http.NewRequest(\"GET\", url, nil)\n\n\treq.Header.Add(\"Authorization\", \"Token <api-key>\")\n\n\tres, _ := http.DefaultClient.Do(req)\n\n\tdefer res.Body.Close()\n\tbody, _ := ioutil.ReadAll(res.Body)\n\n\tfmt.Println(res)\n\tfmt.Println(string(body))\n\n}"
				},
				{
				"lang": "PHP",
				"source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, [\n  CURLOPT_URL => \"https://api.zentry.gg/v1/memories/\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"GET\",\n  CURLOPT_HTTPHEADER => [\n    \"Authorization: Token <api-key>\"\n  ],\n]);\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
				},
				{
				"lang": "Java",
				"source": "HttpResponse<String> response = Unirest.get(\"https://api.zentry.gg/v1/memories/\")\n  .header(\"Authorization\", \"Token <api-key>\")\n  .asString();"
					}
				],
				"x-codegen-request-body-name": "data"
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"MemorySearchInput": {
				"required": [
					"query"
				],
				"type": "object",
				"properties": {
					"query": {
						"title": "Query",
						"minLength": 1,
						"type": "string",
						"description": "The query to search for in the memory."
					},
					"agent_id": {
						"title": "Agent id",
						"type": "string",
						"nullable": true,
						"description": "The agent ID associated with the memory."
					},
					"user_id": {
						"title": "User id",
						"type": "string",
						"nullable": true,
						"description": "The user ID associated with the memory."
					},
					"app_id": {
						"title": "App id",
						"type": "string",
						"nullable": true,
						"description": "The app ID associated with the memory."
					},
					"run_id": {
						"title": "Run id",
						"type": "string",
						"nullable": true,
						"description": "The run ID associated with the memory."
					},
					"metadata": {
						"title": "Metadata",
						"type": "object",
						"properties": {

						},
						"nullable": true,
						"description": "Additional metadata associated with the memory."
					},
					"top_k": {
						"title": "Top K",
						"type": "integer",
						"default": 10,
						"description": "The number of top results to return."
					},
					"fields": {
						"title": "Fields",
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": "A list of field names to include in the response. If not provided, all fields will be returned."
					},
					"rerank": {
						"title": "Rerank",
						"type": "boolean",
						"default": false,
						"description": "Whether to rerank the memories."
					},
					"keyword_search": {
						"title": "Keyword search",
						"type": "boolean",
						"default": false,
						"description": "Whether to search for memories based on keywords."
					},
					"output_format": {
						"title": "Output format",
						"type": "string",
						"nullable": true,
						"default": "v1.0",
						"description": "The search method supports two output formats: `v1.0` (default) and `v1.1`. We recommend using `v1.1` as `v1.0` will be deprecated soon."
					},
					"org_id": {
						"title": "Organization id",
						"type": "string",
						"nullable": true,
						"description": "The unique identifier of the organization associated with the memory."
					},
					"project_id": {
						"title": "Project id",
						"type": "string",
						"nullable": true,
						"description": "The unique identifier of the project associated with the memory."
					},
					"filter_memories": {
						"title": "Filter memories",
						"type": "boolean",
						"default": false,
						"description": "Whether to properly filter the memories according to the input."
					},
					"categories": {
						"title": "Categories",
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": "A list of categories to filter the memories by."
					},
					"only_metadata_based_search": {
						"title": "Only metadata based search",
						"type": "boolean",
						"default": false,
						"description": "Whether to only search for memories based on metadata."
					}
				}
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization",
				"description": "API key authentication. Prefix your Zentry API key with 'Token '. Example: 'Token your_api_key'"
			}
		}
	},
...
}